Partitioned sway config
To use this create a symlink to ~/.config/sway
and make sure the default sway config file performs search like
in fedora.

Fedora sway config has this nice feature to serach for files and
overwritte config for those which have the same name.
This is the old configuration but migrated to the fedora system.